Adds corrections to CellarWines files

+++ src/main/java/andrea_freddi/CAPSTONE_PROJECT/payloads/CellarWineDTO.java
@@ -20,7 +20,7 @@ public record CellarWineDTO(
         Integer purchaseDate,
         BigDecimal purchasePrice,
         BigDecimal askingPrice,
-        Float myScore,
+        Integer myScore,
         UUID userId,
         UUID cellarId,
         UUID wineId
